netology_pd/requirements.txt
django
djangorestframework
psycopg2-binary
pyyaml
requests